ai_planet_assessment/math-mentor-ai/requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM
google-generativeai>=0.3.0

# OCR & ASR
easyocr>=1.7.0
openai-whisper>=20230918
soundfile>=0.12.1

# RAG
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Image & Audio Processing
pillow>=10.0.0
pydub>=0.25.1
numpy>=1.24.0

# Database
# SQLite is built-in to Python
